pyproject.toml
[project]
name = "stock-themes"
version = "0.1.0"
description = "Generate investment themes for stock tickers using public data"
requires-python = ">=3.11"
dependencies = [
    "yfinance>=0.2.30",
    "edgartools>=2.0",
    "openai>=1.30",
    "sentence-transformers>=3.0",
    "torch>=2.0",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
